// src/app/interfaces/user/user.interface.ts
export interface IUser {
    id: string;
    document_number: string;
    dob: string;
    name: string;
    paternal_surname: string;
    maternal_lastname: string;
    email: string;
    phone_number: null;
    username: string;
    status: boolean,
    type_document: string;
    type_gender: string;
    type_role: string;
    created_at: string;
    updated_at: string;
    deleted_at: string;
}